Fortidude: Fix T011: literal-kind
Although it is widely believed that  represents an 8-byte floating point (and indeed, this is the case for most compilers and architectures), there is nothing in the standard to mandate this, and compiler vendors are free to choose any mapping between kind numbers and machine precision. This may lead to surprising results if your code is ported to another machine or compiler.

+++ posti/channel_fft/fft.f90
@@ -276,7 +276,7 @@ USE FFTW3
 IMPLICIT NONE
 !-----------------------------------------------------------------------------------------------------------------------------------
 ! INPUT VARIABLES
-INTEGER(KIND=8),INTENT(IN)       ::plan
+INTEGER(DP),INTENT(IN)           ::plan
 INTEGER,INTENT(IN)               ::nSamples
 INTEGER,INTENT(IN)               ::nSamples_spec
 COMPLEX,INTENT(INOUT)            ::in(nSamples)
