+++ posti/recordpoints/visualize/turbulence/turb.f90
@@ -50,7 +50,7 @@ IMPLICIT NONE
 ! LOCAL VARIABLES
 INTEGER                         :: iSample , iVar,iRP
 INTEGER                         :: nSamples_spec
-INTEGER(KIND=8)                 :: plan
+INTEGER(DP)                     :: plan
 COMPLEX                         :: in(nSamples_out),out(nSamples_out)
 REAL                            :: dt_equi , PI , df
 REAL,ALLOCATABLE                :: vel_spec(:,:,:),velPrim(:,:,:)
